api/requirements.txt
tensorflow=2.18.0
fastapi=0.115.6
uvicorn
pillow
matplotlib
numpy